refactor: extract PrivacyDataExportDiagnosticsSnapshot.swift from PrivacyDataExportTypes.swift
Splits the `PrivacyDataExportDiagnosticsSnapshot` Codable snapshot out
of PrivacyDataExportTypes.swift so that file holds only the
Settings snapshot value.

Byte-identical move of lines 45-57. Both files import Foundation.

PrivacyDataExportTypes.swift: 81 → 67 lines. Tests: 667.
