+++ Sources/BugNarrator/Services/PrivacyDataExportTypes.swift
@@ -42,20 +42,6 @@ struct PrivacyDataExportSettingsSnapshot: Codable, Equatable {
     }
 }
 
-struct PrivacyDataExportDiagnosticsSnapshot: Codable, Equatable {
-    let appName: String
-    let versionDescription: String
-    let macOSVersion: String
-    let architecture: String
-    let activeTranscriptionModel: String
-    let issueExtractionModel: String
-    let logLevel: String
-    let debugModeEnabled: Bool
-    let recentTelemetryEvents: [OperationalTelemetryEvent]
-    let recentDiagnosticsLog: String
-    let exportHistory: [ExportReceipt]
-}
-
 /// A lazy source of stored sessions for export. The exporter pulls sessions one
 /// at a time via `forEach`, so the whole library never has to be materialized in
 /// memory simultaneously (#508). `count` is the number of sessions `forEach` will
